requirements.txt
flask
requests
pyngrok
dialogflow
google-api-core